% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_codominant.R
\name{analyze_codominant_t0}
\alias{analyze_codominant_t0}
\title{Estimate f from codominant marker data assuming theta = 0, i.e.,
assuming that all populations have the same allele frequency at each locus}
\usage{
analyze_codominant_t0(genos, prior_pi, prior_f, ...)
}
\arguments{
\item{genos}{A list in the format returned by read_marker_data()}

\item{prior_pi}{A vector specifying lower and upper limits on pi}

\item{prior_f}{A vector specifying lower and upper limits on f}

\item{...}{Optional arguments passed to \code{rstan::sampling()}}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling()}
}
\description{
Estimate f from codominant marker data assuming theta = 0, i.e.,
assuming that all populations have the same allele frequency at each locus
}
