% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_f.R
\name{analyze_f}
\alias{analyze_f}
\title{Estimate inbreeding coefficients at each population/locus combination
in the data set}
\usage{
analyze_f(genos, ...)
}
\arguments{
\item{genos}{A list in the format retured by \code{read_marker_data()}}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g., iter, chains)}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Estimate inbreeding coefficients at each population/locus combination
in the data set
}
