% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{compare}
\alias{compare}
\title{Compare two models using already calculated estimates of posterior
log likelihoods}
\usage{
compare(model_1, model_2)
}
\arguments{
\item{model_1}{An object of class "\code{loo}" for the first model}

\item{model_2}{An object of class "\code{loo}" for the second model}
}
\value{
A matrix with class "\code{compare.loo}"
}
\description{
Compare two models using already calculated estimates of posterior
log likelihoods
}
