% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{count_genos}
\alias{count_genos}
\title{Count the genotypes at a single locus in a single population}
\usage{
count_genos(x)
}
\arguments{
\item{x}{Vector of genotypes at a single locus in a single population
\itemize{
\item 0, 1, 2 for codominant markers
\item 0, 1 for dominant markers
}}
}
\value{
A table with the number of each genotype
}
\description{
Count the genotypes at a single locus in a single population
}
