% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose_bulk_ess}
\alias{diagnose_bulk_ess}
\title{Report bulk ESS for parameters exceeding a specified threshold}
\usage{
diagnose_bulk_ess(fit, threshold = 400)
}
\arguments{
\item{fit}{The model results for which to check bulk ESS}

\item{threshold}{The threshold below which to report parameters}
}
\description{
Report bulk ESS for parameters exceeding a specified threshold
}
