% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose_tail_ess}
\alias{diagnose_tail_ess}
\title{Report tail ESS for parameters exceeding a specified threshold}
\usage{
diagnose_tail_ess(fit, threshold = 400)
}
\arguments{
\item{fit}{The model results for which to check tail ESS}

\item{threshold}{The threshold below which to report parameters}
}
\description{
Report tail ESS for parameters exceeding a specified threshold
}
