% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{energy_plot}
\alias{energy_plot}
\title{Plot energy versus specified parameters for one chain}
\usage{
energy_plot(chain, fit, pars)
}
\arguments{
\item{chain}{The number of the chain to plot}

\item{fit}{The model results from which to extract results}

\item{pars}{A vector of parameters to plot against energy}
}
\description{
Plot energy versus specified parameters for one chain
}
