% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_estimate.R
\name{f_estimate}
\alias{f_estimate}
\title{Estimate inbreeding coefficient at one locus with two alleles}
\usage{
f_estimate(n)
}
\arguments{
\item{n}{Numeric vector of genotype counts}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g., iter, chains)}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Estimate inbreeding coefficient at one locus with two alleles
}
