% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{initialize_locus_pop}
\alias{initialize_locus_pop}
\title{Initialize locus-specific thetas
Note: Depends on prior_theta, prior_alpha, and N_loci being defined in this
namespace. Implicitly depends on prior_pi and prior_f being defined in this
namespace.}
\usage{
initialize_locus_pop()
}
\value{
Initial values for all internal parameters
}
\description{
Initialize locus-specific thetas
Note: Depends on prior_theta, prior_alpha, and N_loci being defined in this
namespace. Implicitly depends on prior_pi and prior_f being defined in this
namespace.
}
