% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{logit_prior}
\alias{logit_prior}
\title{Set mean and standard deviation on the logit scale given lower and upper
bound}
\usage{
logit_prior(prior)
}
\arguments{
\item{prior}{A list with two components: mean and upper}
}
\value{
A vector with element 1 = mean and element 2 = sd
}
\description{
Set mean and standard deviation on the logit scale given lower and upper
bound
}
