% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo}
\alias{loo}
\title{Leave-one-out cross validation for Hickory models}
\usage{
loo(fit, ...)
}
\arguments{
\item{fit}{A stanfit object returned from \code{analyze_codominant()} or
\code{analyze_dominant()}}

\item{...}{Optional parameters passed to \code{loo::loo()}}
}
\value{
A named list from \code{loo::loo()}.
}
\description{
Use leave-one-out cross validation to estimate the expected log pointwise
predictive density (\code{elpd_loo}), the effective number of parameters,
(\code{p_loo}), and the LOO information criterion (\code{looic}). See the
documentation for \code{loo()} in the \code{loo} package for details.
}
