% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_marker_data}
\alias{read_marker_data}
\title{Read marker data from a CSV file}
\usage{
read_marker_data(filename)
}
\arguments{
\item{filename}{Full path name to a CSV file containing marker data}
}
\value{
A list with components
N_pops for the number of populations in the sample
N_loci for the number of genotypes in the sample
n A N_pops x N_loci x 3 array of genotype counts
N A N_pops x N_loci matrix of sample sizes
}
\description{
Read marker data from a CSV file
}
