% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Summarize locus-, population-, or both specific estimates of theta}
\usage{
report(fit, genos, locus = FALSE, pop = FALSE, alpha = 0.05)
}
\arguments{
\item{fit}{The model results to summarize}

\item{genos}{The original genotype data}

\item{locus}{If TRUE report locus-specific effects}

\item{pop}{If TRUE report population-specific effects}

\item{alpha}{1 - alpha central credible interval}
}
\description{
Summarize locus-, population-, or both specific estimates of theta
}
