% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_}
\alias{report_}
\title{Summarize estimates. Not intended for direct use. Use summarize()}
\usage{
report_(labels, theta, alpha = 0.05)
}
\arguments{
\item{labels}{Labels for each item in the summary}

\item{theta}{Individual estimates of theta for each item}

\item{alpha}{1 - alpha central credible interval}
}
\description{
Summarize estimates. Not intended for direct use. Use summarize()
}
