% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{report_outliers}
\alias{report_outliers}
\title{Report locus or population outliers}
\usage{
report_outliers(fit, genos, alpha = 0.05, locus = TRUE, pop = TRUE)
}
\arguments{
\item{fit}{The model results to check for outliers}

\item{genos}{The original genotype data}

\item{alpha}{Use 1 - alpha central credible interval to identify outliers}

\item{locus}{If TRUE report locus outliers}

\item{pop}{If TRUE report population outliers

Locus or population specific estimates of theta are reported if the credible interval for the difference
between those estimates and the overall estimate of theta does not overlap 0. The report lists the
locus or population identified as an outlier, the mean difference between the specific and overall
estimate and the (alpha/2, 1 - alpha/2) credible interval for the difference.}
}
\description{
Report locus or population outliers
}
