% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{report_outliers_}
\alias{report_outliers_}
\title{Report outliers. Not intended for direct use. Use report_outliers().}
\usage{
report_outliers_(labels, theta, theta_l, alpha)
}
\arguments{
\item{labels}{Labels for each item in the report}

\item{theta}{Central estimate}

\item{theta_l}{Individual estimate}

\item{alpha}{1 - alpha central credible interval}
}
\description{
Report outliers. Not intended for direct use. Use report_outliers().
}
