% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{set_alpha}
\alias{set_alpha}
\title{Put alpha ("tightness" of locus-specific theta) in global namespace
so that it is available to \code{initialize_loc_pop()}.}
\usage{
set_alpha(alpha)
}
\arguments{
\item{alpha}{The value of alpha to store}
}
\value{
None
}
\description{
Put alpha ("tightness" of locus-specific theta) in global namespace
so that it is available to \code{initialize_loc_pop()}.
}
