% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{set_priors}
\alias{set_priors}
\title{Put priors in Hickory namespace so that they are available to
\code{initialize_chains()}.}
\usage{
set_priors(prior_pi, prior_f, prior_theta, N_loci, N_pops)
}
\arguments{
\item{prior_pi}{A vector with the mean and upper bound on pi}

\item{prior_f}{A vector with the mean and upper bound on f}

\item{prior_theta}{A vector with the mean and upper bound on theta}

\item{N_loci}{The number of loci in the sample}

\item{N_pops}{The number of populations in the sample}
}
\value{
None
}
\description{
Put priors in Hickory namespace so that they are available to
\code{initialize_chains()}.
}
