% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_f.R
\name{summarize_f}
\alias{summarize_f}
\title{Summarize the results of \code{analyze_f()}}
\usage{
summarize_f(fit, genos, prob = 0.05)
}
\arguments{
\item{fit}{A stanfit object returned by \code{analyze_f()}}

\item{genos}{The genotype counts that were analyzed}

\item{prob}{Sets the credible interval to flag population locus combinations,
i.e., those where the central 1-\code{prob} credible interval does not overlap 0}
}
\value{
No values returned
}
\description{
Summarize the results of \code{analyze_f()}
}
