% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_codominant.R
\name{analyze_codominant}
\alias{analyze_codominant}
\title{Estimate f and theta from codominant marker data}
\usage{
analyze_codominant(
  genos,
  prior_pi = list(lower = 0.1, upper = 0.9),
  prior_f = list(lower = 0.01, upper = 0.2),
  prior_theta = list(lower = 0.01, upper = 0.2),
  ...
)
}
\arguments{
\item{genos}{A list in the format returned by read_marker_data()}

\item{prior_pi}{A vector specifying lower and upper limits on pi}

\item{prior_f}{A vector specifying lower and upper limits on f}

\item{prior_theta}{A vector specifying lower and upper limits on theta}

\item{...}{Optional arguments passed to \code{rstan::sampling()}}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling()}
}
\description{
Estimate f and theta from codominant marker data
}
