% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_f.R
\name{analyze_f}
\alias{analyze_f}
\title{Estimate inbreeding coefficients at each population/locus combination
in the data set}
\usage{
analyze_f(genos, prob = 0.05, ...)
}
\arguments{
\item{genos}{A list in the format retured by \code{read_marker_data()}}

\item{prob}{Sets the credible interval to flag population locus combinations,
i.e., those where the central 1-\code{prob} credible interval does not overlap 0}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g., iter, chains)}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
The summary printed before the function exits identifies the
loci for which the central (1-\code{prob}) credible interval for f does not
include 0.
}
