% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hickory-package.R
\docType{package}
\name{Hickory-package}
\alias{Hickory-package}
\alias{Hickory}
\title{The 'Hickory' package.}
\description{
Hickory provides functions to read genotype data from simple
CSV files and to estimate the within population inbreeding coefficient, f,
Wright's Fst, theta.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
