% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{initialize_chains}
\alias{initialize_chains}
\title{Initialize pi, f, and theta based on priors
Note: Depends on prior_pi, prior_f, and prior_theta being defined in
this namespace}
\usage{
initialize_chains()
}
\value{
None
}
\description{
Initialize pi, f, and theta based on priors
Note: Depends on prior_pi, prior_f, and prior_theta being defined in
this namespace
}
